Create Day 1 - Band Name Generator
It is day 1, the start of a wonderful journey, although I know python pretty well I think practicing it always will be great as I am a beginner. today I made a small program using an input function that uses your city and pet name to generate a band name for you.